requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
streamlit>=1.32.0

requests>=2.31.0
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0

pandas>=2.1.1
numpy>=1.26.0
plotly>=5.18.0

pydantic>=2.4.2

langchain-google-genai>=0.1.0
langchain
langchain-core>=0.2.0
langchain-openai>=0.1.0
langchain-chroma>=0.1.0
chromadb>=0.5.0
openai>=1.30.0

typing-extensions>=4.9.0